requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.36
asyncpg==0.30.0      # Для работы с PostgreSQL асинхронно
alembic==1.14.0
cachetools==5.5.0
PyJWT==2.9.0
passlib==1.7.4
pytest==8.3.3
databases==0.9.0
email-validator==2.2.0
python-multipart==0.0.17
httpx==0.27.2
pytest-asyncio==0.24.0